Move versions JSON to test source set
