+++ .github/renovate.json5
@@ -14,11 +14,11 @@
       // Matches AGP versions in files with custom patterns
       customType: "regex",
       fileMatch: [
-        "src/main/resources/testedVersions\\.json",
+        "src/test/resources/versions\\.json",
         "gradle\\.properties",
       ],
       "matchStrings": [
-        // In testedVersions.json, matches keys with a list value, e.g. "8.9.0": [
+        // In versions.json, matches keys with a list value, e.g. "8.9.0": [
         "\\s+?\"(?<currentValue>\\S+?)\": \\[",
         // In gradle.properties, matches a specific property's value
         "org\\.gradle\\.android\\.latestKnownAgpVersion=(?<currentValue>\\S+?)(?:\\s|$)",
@@ -30,14 +30,14 @@
     },
   ],
   // Ensure patches to older minors are opened, even if a newer minor is available
-  // In testedVersions.json, means older minors will still be checked for a newer patch
+  // In versions.json, means older minors will still be checked for a newer patch
   "separateMinorPatch": true,
   "packageRules": [
     {
-      // In testedVersions.json, disable bumps of major/minor, which should be added as new properties
+      // In versions.json, disable bumps of major/minor, which should be added as new properties
       "matchDepNames": ["com.android.tools.build:gradle"],
       "matchUpdateTypes": ["major", "minor"],
-      "matchFileNames": ["src/main/resources/testedVersions\\.json"],
+      "matchFileNames": ["src/test/resources/versions\\.json"],
       "enabled": false,
     },
     {
@@ -49,7 +49,7 @@
     {
       // Group changes about AGP versions
       "matchFileNames": [
-        "src/main/resources/testedVersions\\.json",
+        "src/test/resources/versions\\.json",
         "gradle\\.properties",
       ],
       "commitMessageTopic": "tested AGP versions",
