+++ src/test/groovy/org/gradle/android/VersionsTest.groovy
@@ -3,11 +3,11 @@ package org.gradle.android
 import spock.lang.Specification
 
 class VersionsTest extends Specification {
-    def "earliest supported version includes alpha and beta builds"() {
-        def alphaOfEarliest = alphaVersionOf(Versions.TESTED_ANDROID_VERSIONS.min())
+    def "earliest tested version corresponds to minimum supported version"() {
+        def alphaOfEarliest = alphaVersionOf(TestVersions.TESTED_ANDROID_VERSIONS.min())
 
         expect:
-        alphaOfEarliest >= Versions.earliestSupportedAndroidVersion()
+        alphaOfEarliest == Versions.MINIMUM_ANDROID_VERSION
     }
 
     static VersionNumber alphaVersionOf(VersionNumber versionNumber) {
